feat: DANGER dark / light mode
