+++ src/admin/adminComponents/AdminHeader.jsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import AdminNavbar from './AdminNavbar';
-import theme from '../../utils/theme';
+
+import { Box } from '@chakra-ui/react';
 
 export default function AdminHeader() {
   return (
-    <header
-      className='page-header'
-      style={{ background: theme.colors.neutral.bg, height: 'auto', zIndex: '1', paddingBottom: '30px' }}
-    >
+    <Box bg='neutral.bg' className='page-header' style={{ height: 'auto', zIndex: '1', paddingBottom: '30px' }}>
       <AdminNavbar />
-    </header>
+    </Box>
   );
 }
