+++ src/user/userComponents/AccountSettingsPage/SettingsMenu.jsx
@@ -1,12 +1,13 @@
-import { useState, useCallback, useEffect } from 'react';
+import { useState, useCallback, useEffect, useContext } from 'react';
 import { Box, Flex, Heading, Text } from '@chakra-ui/react';
 import UserAccount from '../../../assets/svg/UserAccount';
 import Address from '../../../assets/svg/Address';
 import Security from '../../../assets/svg/Security';
-import theme from '../../../utils/theme';
+
 import { Link, useLocation } from 'react-router-dom';
 import React from 'react';
 import PaymentMethod from '../../../assets/svg/PaymentMethod';
+
 function SettingItem({ element, isSelected, onItemSelected }) {
   const [isHovered, setIsHovered] = useState(false);
   const Icon = element.icon;
@@ -43,7 +44,7 @@ function SettingItem({ element, isSelected, onItemSelected }) {
   const borderColor = isSelected ? 'primary.default' : isHovered ? 'primary.default' : 'neutral.grayLightest';
   const bgColor = isSelected ? 'primary.default' : isHovered ? 'primary.lightest' : 'neutral.white';
   const textColor = isSelected ? 'primary.default' : 'neutral.black';
-  const iconColor = isSelected ? theme.colors.neutral.white : theme.colors.neutral.black;
+  const iconColor = isSelected ? 'white' : 'black';
 
   return (
     <Box
